organize options into a directory
