+++ cs/cli/vw_arguments.h
@@ -6,7 +6,7 @@
 
 #include <msclr\marshal_cppstd.h>
 #include "vw.h"
-#include "options_serializer_boost_po.h"
+#include "cli_options_serializer.h"
 #include <algorithm>
 
 using namespace System;
@@ -47,7 +47,7 @@ private:
         m_regressors->Add(gcnew String(r.c_str()));
     }
 
-    VW::config::options_serializer_boost_po serializer;
+    VW::config::cli_options_serializer serializer;
     for (auto const& option : options->get_all_options())
     {
       if (options->was_supplied(option->m_name))
