+++ vowpalwabbit/cached_learner.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "global_data.h"  // to get vw struct
-#include "options.h"      // to get options_i
+#include "config/options.h"      // to get options_i
 
 namespace VW
 {
