+++ vowpalwabbit/config/help_formatter.h
@@ -1,0 +1,28 @@
+// Copyright (c) by respective owners including Yahoo!, Microsoft, and
+// individual contributors. All rights reserved. Released under a BSD (revised)
+// license as described in the file LICENSE.
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+namespace VW
+{
+namespace config
+{
+
+struct option_group_definition;
+struct options_i;
+
+std::vector<option_group_definition> remove_disabled_necessary_options(
+    options_i& options, const std::vector<option_group_definition>& groups);
+
+struct help_formatter
+{
+  virtual std::string format_help(const std::vector<option_group_definition>& groups) = 0;
+  virtual ~help_formatter() = default;
+};
+
+}  // namespace config
+}  // namespace VW
