+++ vowpalwabbit/offset_tree.h
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "learner.h"
-#include "options.h"
+#include "config/options.h"
 
 namespace VW
 {
