+++ vowpalwabbit/pmf_to_pdf.h
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "learner.h"
-#include "options.h"
+#include "config/options.h"
 #include "cb.h"
 #include "action_score.h"
